debug: move code requring internal access to repository package
